Splash screen added

+++ app/src/main/java/org/wit/carapp/models/CarModel.kt
@@ -7,6 +7,6 @@ import kotlinx.android.parcel.Parcelize
 data class CarModel(var id: Long = 0,
                           var make: String = "",
                           var model: String = "",
-                          var year: Short = 0,
+                          var year: Int = 0,
                           var engine: Double = 0.0,
                           var image: String = "") : Parcelable
